+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/MapView.java
@@ -1623,7 +1623,7 @@ public class MapView extends FrameLayout {
         return direction;
     }
 
-    void setBearing(float bearing) {
+    void setBearing(double bearing) {
         if (destroyed) {
             return;
         }
@@ -1631,7 +1631,7 @@ public class MapView extends FrameLayout {
         nativeMapView.setBearing(bearing);
     }
 
-    void setBearing(float bearing, long duration) {
+    void setBearing(double bearing, long duration) {
         if (destroyed) {
             return;
         }
